1. Created a new project named BattleSystem 2. Within BattleSystem, imported all JavaFX files related to battle system UI (not operational yet)
Note: Currently, the system can only open up 1 scene at a time. The difference between JavaFX files and their respective controllers really are just the number of allies/enemies on screen.
If you want to change the screen that is shown, use the HelloApplication.java file and change the fxml file name to the one you want to call.

Warnings all about removing "default property tag", or claiming that functions arte not being used. All of them should be harmless.